Data: Store tail storage node link. Do not store allocator in node. Store array pool setup in data instead of 2 links to node and store pools. ushort used in link node. Issue with remove when all nodes are full fixed.

+++ Konsarpoo.Tests/MapTest.cs
@@ -15,7 +15,7 @@ namespace Konsarpoo.Collections.Tests
     [TestFixture(1024, AllocatorType.Pool, 0)]
     public class MapTest : BaseTest
     {
-        public MapTest(int? maxSizeOfArrayBucket, AllocatorType allocatorType, int gcLen) : base(maxSizeOfArrayBucket, allocatorType, gcLen)
+        public MapTest(int? maxSizeOfArrayBucket, AllocatorType allocatorType, int gcLen) : base(maxSizeOfArrayBucket, allocatorType, (int)gcLen)
         {
         }
